requirements.txt
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
Flask==1.1.1
Flask-Testing==0.7.1
gunicorn==20.0.4
idna==2.8
importlib-metadata==1.3.0
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
more-itertools==8.0.2
packaging==19.2
pluggy==0.13.1
py==1.8.0
pyparsing==2.4.5
pytest==5.3.2
python-dotenv==0.10.3
requests==2.22.0
selenium==3.141.0
six==1.13.0
urllib3==1.25.7
wcwidth==0.1.7
Werkzeug==0.16.0
zipp==0.6.0
